refactor: update and enhance examples

+++ examples/aws-schedule-manager/src/api/endpoints/delete-schedule.ts
@@ -3,6 +3,10 @@ import type { Service } from '@ez4/common';
 import type { Http } from '@ez4/gateway';
 import type { Api } from '../../api.js';
 
+import { HttpBadRequestError } from '@ez4/gateway';
+
+import { deleteEvent } from '../repository.js';
+
 /**
  * Delete schedule request.
  */
@@ -16,11 +20,7 @@ export declare class DeleteScheduleRequest implements Http.Request {
  * Delete schedule response.
  */
 export declare class DeleteScheduleResponse implements Http.Response {
-  status: 200;
-
-  body: {
-    message: string;
-  };
+  status: 204;
 }
 
 /**
@@ -30,15 +30,20 @@ export async function deleteScheduleHandler(
   request: DeleteScheduleRequest,
   context: Service.Context<Api>
 ): Promise<DeleteScheduleResponse> {
+  const { eventDb, eventScheduler } = context;
   const { scheduleId } = request.parameters;
-  const { eventScheduler } = context;
+
+  const exists = await eventScheduler.getEvent(scheduleId);
+
+  if (!exists) {
+    throw new HttpBadRequestError(`Schedule doesn't exist`);
+  }
 
   await eventScheduler.deleteEvent(scheduleId);
 
+  await deleteEvent(eventDb, scheduleId);
+
   return {
-    status: 200,
-    body: {
-      message: 'success'
-    }
+    status: 204
   };
 }
